h-1

+++ src/components/user/transaction/Checkout.js
@@ -2,7 +2,8 @@ import React, { Component } from 'react'
 import axios from 'axios'
 import { connect } from 'react-redux'
 import { Redirect } from 'react-router-dom'
-import { Button, Card, CardBody, CardText, CardTitle, CardHeader } from 'reactstrap';
+import { Button, Card, CardBody, CardText, CardTitle } from 'reactstrap';
+import Swal from 'sweetalert2'
 
 import Header from '../header/Header'
 import Footer from '../../Footer'
@@ -10,19 +11,30 @@ import Footer from '../../Footer'
 class Checkout extends Component {
    state = {
       cart: [],
+      pending: [],
+      cancel: [],
       user_cart: [],
       product: [],
-      redirect: false
+      redirect: false,
+      redirectCancel: false
    }
 
    componentDidMount() {
       this.getCart()
       this.getProduct()
+      this.getCheckout()
+      this.getCheckoutCancel()
    }
 
    setRedirect = () => {
       this.setState({
          redirect: true
+      })
+   }
+
+   setRedirectCancel = () => {
+      this.setState({
+         redirectCancel: true
       })
    }
 
@@ -51,6 +63,24 @@ class Checkout extends Component {
 
       this.setState({ user_cart: filterCart })
       console.log(filterCart)
+   }
+
+   getCheckout = () => {
+      let user_id = this.props.match.params.user_id
+
+      axios.get(`http://localhost:2019/checkoutpending/${user_id}`)
+         .then(res => {
+            this.setState({ pending: res.data })
+         })
+   }
+
+   getCheckoutCancel = () => {
+      let user_id = this.props.match.params.user_id
+
+      axios.get(`http://localhost:2019/checkoutcancel/${user_id}`)
+         .then(res => {
+            this.setState({ cancel: res.data })
+         })
    }
 
    renderCart = () => {
@@ -89,56 +119,92 @@ class Checkout extends Component {
    }
 
    onCheckout = async () => {
-      const total_price = this.onTotal()
-      const user_id = this.props.user.id
+      if (this.state.pending.length > 0) {
+         Swal.fire({
+            type: 'error',
+            title: 'Access Denied',
+            text: 'You have pending transaction, please complete it first before do another transaction!',
+         })
+         this.setRedirect()
+      } else if (this.state.cancel.length > 0) {
+         Swal.fire({
+            type: 'error',
+            title: 'Transaction Denied',
+            text: 'Your transaction has been denied!',
+         })
+         this.setRedirectCancel()
+      } else if (this.state.user_cart.length === 0) {
+         Swal.fire({
+            type: 'error',
+            title: 'Empty Cart',
+            text: 'Your cart is empty!',
+         })
+      } else {
+
+         const total_price = this.onTotal()
+         const user_id = this.props.user.id
+         const order_status = 'Transaction Pending'
 
 
-      console.log(total_price)
+         console.log(total_price)
 
-      const order = await axios.post(`http://localhost:2019/checkout`,
-         {
-            user_id,
-            total_price
+         const order = await axios.post(`http://localhost:2019/checkout`,
+            {
+               user_id,
+               total_price,
+               order_status
+            })
+
+         console.log(order)
+
+         let arrCart = []
+         let carts = this.state.user_cart
+
+         console.log(carts)
+
+         for (let i = 0; i < carts.length; i++) {
+            arrCart.push([carts[i].product_id, carts[i].quantity, order.data[0].id])
+         }
+
+         console.log(arrCart)
+
+         const orderDetail = await axios.post(`http://localhost:2019/orderdetail`, { arrCart })
+
+         console.log(orderDetail)
+
+         await axios.delete(`http://localhost:2019/carts/${this.props.user.id}`)
+
+         Swal.fire({
+            type: 'success',
+            title: 'Checkout Success!',
+            text: 'Order has been created, please upload payment receipt',
          })
 
-      console.log(order)
-
-      let arrCart = []
-      let carts = this.state.user_cart
-
-      console.log(carts)
-
-      for (let i = 0; i < carts.length; i++) {
-         arrCart.push([carts[i].product_id, carts[i].quantity, order.data[0].id])
+         this.setRedirect()
       }
-
-      console.log(arrCart)
-
-      const orderDetail = await axios.post(`http://localhost:2019/orderdetail`, { arrCart })
-
-      console.log(orderDetail)
-
-      await axios.delete(`http://localhost:2019/carts/${this.props.user.id}`)
-
-      this.setRedirect()
    }
 
    renderRedirect = () => {
       if (this.state.redirect) {
-         return <Redirect to='/' />
+         return <Redirect to={`/confirmation/${this.props.user.id}`} />
       }
    }
 
+   renderRedirectCancel = () => {
+      if (this.state.redirectCancel) {
+         return <Redirect to={`/order/${this.props.user.id}`} />
+      }
+   }
 
    render() {
       console.log(this.onTotal)
       return (
-         <div style={{backgroundImage: "url(http://www.hdcarwallpapers.com/walls/2019_ssc_tuatara_4k-HD.jpg)", backgroundSize: 'cover', backgroundPosition: '78%'}}>
+         <div style={{ backgroundImage: "url(http://www.hdcarwallpapers.com/walls/2019_ssc_tuatara_4k-HD.jpg)", backgroundSize: 'cover', backgroundPosition: '78%' }}>
             <Header />
-            <div className='container mb-5' style={{ marginTop: 88  }}>
+            <div className='container mb-5' style={{ marginTop: 88 }}>
                <div className="w-50 mx-auto" style={{ opacity: 0.9 }}>
                   <Card body className='mb-3'>
-                  <CardText className='text-center display-4'>Cart</CardText>
+                     <CardText className='text-center display-4'>Cart</CardText>
                   </Card>
                   <Card body>
                      <CardBody>
@@ -153,6 +219,7 @@ class Checkout extends Component {
                         <div>
                            <Button className="btn btn-danger text-white btn-md btn-block" onClick={() => this.onCheckout()}>Continue Order</Button>
                            {this.renderRedirect()}
+                           {this.renderRedirectCancel()}
                         </div>
                      </CardBody>
                   </Card>
